+++ src/routes/+layout.svelte
@@ -1,10 +1,55 @@
 <script lang="ts">
   import {
+    Col,
+    Collapse,
     Container,
+    Dropdown,
+    DropdownItem,
+    DropdownMenu,
+    DropdownToggle,
+    Icon,
+    Image,
+    Nav,
+    Navbar,
+    NavbarBrand,
+    NavItem,
+    NavLink,
+    NavbarToggler,
+    Row,
     Styles
   } from 'sveltestrap';
+  import { writable } from 'svelte/store';
   import { setContext } from 'svelte';
-  import {writable } from 'svelte/store';
+  
+    // Display mode logic
+    interface DisplayMode {
+    buttonText: string;
+    dropdownText: string;
+  }
+  let displayMode = writable("app");
+  setContext('displayMode', displayMode);
+  let displayModes:Record<string, DisplayMode> = {
+    app: {
+      buttonText: "App Interpretation",
+      dropdownText: "Display app's interpretation of resources"
+    },
+    text: {
+      buttonText: "Generated Text",
+      dropdownText: "Display text generated by IPS source"
+    }
+  }
+
+  let displayModeText:string;
+  $: {
+    if ($displayMode) {
+      displayModeText = displayModes[$displayMode].buttonText;
+    }
+  }
+
+  function setMode(mode:string) {
+    $displayMode = mode;
+  }
+  // End display mode logic
 
   const MODE_KEY = 'demo_mode';
   let mode = writable('normal');
@@ -20,13 +65,15 @@
   document.addEventListener('click', (event) => {
     // Ignore clicks on the navbar toggler
     if (event.target instanceof Element) {
-      if (event.target?.className?.includes('navbar-toggler')) return;
+      if (event.target?.className?.includes('dropdown-toggle')) return;
       // Ignore clicks on the dropdown toggle menu items
-      if (event.target?.className?.includes('nav-link') && event.target?.className?.includes('dropdown-toggle')) {
+      if (event.target?.className?.includes('nav-link') && event.target?.className?.includes('dropdown-toggle')
+          || event.target?.className?.includes('navbar-toggler')
+          || event.target?.className?.includes('navbar-toggler-icon')) {
         navOpening = true;
         setTimeout(() => {
           navOpening = false;
-        }, 100);
+        }, 1000);
         return;
       }
     }
@@ -49,61 +96,117 @@
 
   setContext('mode', mode);
 
+  function handleUpdate(event: any) {
+    $isOpen = event.detail.isOpen;
+  }
+
 </script>
 
 <Container class="main" fluid>
   <Styles />
-  <slot />
+  <Navbar class="sticky-top" color="light" light expand="sm" style="border-bottom: 1px solid rgb(204, 204, 204);">
+    <NavbarBrand class="p-0" href="https://international-patient-summary.net" target="_blank">
+      <Row>
+        <Col>
+          <Image
+            id="nav-image"
+            alt="International Patient Summary"
+            src="/img/ips-logo.png"
+            style="max-height: 75px"
+          />
+        </Col>
+        <Col style="vertical-align:middle" class="d-none d-md-block align-items-center">
+          <div class="nav-text mt-1">
+            <span>International Patient</span>
+            <p style="margin: 0px; padding: 0px: line-height: 0;"></p>
+            <span>Summary Viewer</span>
+          </div>
+        </Col>
+      </Row>
+    </NavbarBrand>
+    <NavbarToggler on:click={() => ($isOpen = !$isOpen)} />
+    <Collapse class="flex-row d-flex justify-content-end" isOpen={$isOpen} navbar expand="sm" on:update={handleUpdate}>
+      <Row class="d-flex justify-content-end align-items-center">
+        Displaying FHIR Resources Using:
+        <Col style="width: fit-content">
+          <Dropdown>
+            <DropdownToggle color="secondary" caret>
+              {displayModeText}
+            </DropdownToggle>
+            <DropdownMenu>
+              {#each Object.entries(displayModes) as [mode, {buttonText, dropdownText}]}
+                <DropdownItem
+                  on:click={() => {
+                    setMode(mode);
+                }}>
+                  <Row class="mr-0" style="min-width:340px">
+                    <Col class="d-flex justify-content-start align-items-center pe-0">
+                      {dropdownText}
+                    </Col>
+                    <Col class="d-flex justify-content-end ps-0">
+                      {#if mode === $displayMode} <Icon name="check" class="text-success fs-5"/>{/if}
+                    </Col>
+                  </Row>
+                </DropdownItem>
+              {/each}
+            </DropdownMenu>
+          </Dropdown>
+        </Col>
+      </Row>
+    </Collapse>
+  </Navbar>
+  <Row 
+    style="border-bottom: 1px solid rgb(204, 204, 204);"
+    class="d-flex justify-content-between align-items-center px-2 py-3"
+  >
+  <h1>International Patient Summary (IPS) Viewer for Connectathon</h1>
+  <div>
+    Links to <a href="https://hl7.org/fhir/uv/ips/">published Implementation Guide</a> and the <a
+      href="http://build.fhir.org/ig/HL7/fhir-ips/">latest CI build</a>
+  </div>
+  <div>
+    Please note that this tool is an open-source project under development. It only renders the following sections of
+    IPS bundles: Advance Directives, Allergies, Immunizations, Medications, Problems, Results, and Social History. Rendering of FHIR resources
+    may be incomplete and using the narrative may be necessary in some circumstances. 
+  </div>
+  </Row>
+  
+  <Row class="main-row">
+    <Col>
+      <slot />
+    </Col>
+  </Row>
+  <Row>
+    <Col style="margin-top: 20px; padding: 20px; border-top: 1px solid rgb(204, 204, 204);" >
+      <footer>
+        <lead>Originally created by Alejandro Lopez Osornio, Diego Kaminker and Fernando Campos. Modified 2021-2023 by John
+          D'Amore</lead>
+        <br />
+        <lead>Based on <a href="https://salud-ar.github.io/IPS-Argentina/demo.html">prior work</a> from <a
+            href="https://github.com/SALUD-AR/IPS-Argentina">this repository</a></lead>
+        <br />
+        <lead>Licensed according to Apache 2.0. See <a href="https://github.com/jddamore/IPSviewer">current code
+            repository</a> for details</lead>
+        <br />
+        <lead>Hosted by <a href="https://www.moreinformatics.com">More Informatics, Inc.</a></lead>
+      </footer>
+    </Col>
+  </Row>
 </Container>
 
 <style>
-  :global(#nav-image) {
-    width: 240px;
-    -webkit-transition: all 0.06s linear;
-    -moz-transition: all 0.06s linear;
-    -o-transition: all 0.06s linear;
-    transition: all 0.06s linear;
-  }
-  :global(.nav-text) {
-    font-size:medium;
-    -webkit-transition: all 0.06s linear;
-    -moz-transition: all 0.06s linear;
-    -o-transition: all 0.06s linear;
-    transition: all 0.06s linear;
-  }
-  :global(.nav-link.scrolling) {
-    padding-top: 0rem !important;
-    padding-bottom: 0.25rem !important;
-  }
-  :global(#nav-image.scrolling) {
-    width: 160px !important;
-    margin-left: 10px;
-  }
-  :global(.nav-text.scrolling)  {
-    font-size: xx-small;
-    color: #000; /* Fallback for older browsers */
-    color: rgba(0, 0, 0, 0.0);
-  }
-  :global(.navbar.scrolling) {
-    padding: 0px !important;
-  }
-
-  :global(.main-row) {
-    flex-grow: 1;
-  }
-
   :global(div.container-fluid.main) {
     min-height: 100%;
     margin-right: auto;
     margin-left: auto;
-    max-width: 800px;
+    max-width: 100%;
     display: flex;
     flex-direction: column;
   }
   :global(html, body) {
     height: 100%;
   }
-  :global(.navbar .container-fluid) {
-    padding: 0px;
+  :global(.main-row) {
+    flex-grow: 1;
   }
 </style>
